doc-02 + feat-04: Validação de reservas

+++ src/controllers/HouseController.js
@@ -63,9 +63,9 @@ class HouseController {
         const user = await User.findById(user_id);
         const houses = await House.findById(house_id);
 
-        //if (String(user._id) !== String(houses.user)) {
-        //    return res.status(401).json({ error: 'Não Autorizado' })
-        //}
+        if (String(user._id) !== String(houses.user)) {
+            return res.status(401).json({ error: 'Não Autorizado' })
+        }
 
         await House.findByIdAndDelete({ _id:house_id});
